Add test for mysql metrics
Also moved tests for openstack metrics in separate file
